requirements.txt

Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
marshmallow-sqlalchemy==0.23.1
